requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.2

# HTTP and Multipart
requests==2.31.0
python-multipart==0.0.6

# Machine Learning and Vector Search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy>=1.24.3
torch>=2.0.0
transformers>=4.30.0

# Text Processing and Utilities
chardet==5.2.0
huggingface-hub>=0.19.0
tqdm>=4.65.0
scikit-learn>=1.2.2

# Logging and Monitoring
logging>=0.5.1.2

# Date and Time
python-dateutil>=2.8.2